Move Observer/Observable to utils
