+++ cmake/NeuronFileLists.cmake
@@ -421,7 +421,7 @@ set(NMODL_FILES_LIST
     units.cpp
     version.cpp)
 
-set(IVOS_FILES_LIST observe.cpp resource.cpp)
+set(IVOS_FILES_LIST resource.cpp)
 
 set(MPI_DYNAMIC_INCLUDE nrnmpi_dynam.h nrnmpi_dynam_cinc nrnmpi_dynam_wrappers.inc)
