+++ src/ivos/InterViews/_defines.h
@@ -185,9 +185,6 @@
 #define OLDialogKit _lib_iv(OLDialogKit)
 #define OLKit _lib_iv(OLKit)
 #define OLKitImpl _lib_iv(OLKitImpl)
-#define Observable _lib_iv(Observable)
-#define Observer _lib_iv(Observer)
-#define ObserverList _lib_iv(ObserverList)
 #define OptionDesc _lib_iv(OptionDesc)
 #define PSFont _lib_iv(PSFont)
 #define PSFontImpl _lib_iv(PSFontImpl)
