--- src/ivos/observe.cpp
@@ -1,57 +1,0 @@
-#ifdef HAVE_CONFIG_H
-#include <../../nrnconf.h>
-#endif
-/*
- * Copyright (c) 1992 Stanford University
- * Copyright (c) 1992 Silicon Graphics, Inc.
- *
- * Permission to use, copy, modify, distribute, and sell this software and 
- * its documentation for any purpose is hereby granted without fee, provided
- * that (i) the above copyright notices and this permission notice appear in
- * all copies of the software and related documentation, and (ii) the names of
- * Stanford and Silicon Graphics may not be used in any advertising or
- * publicity relating to the software without the specific, prior written
- * permission of Stanford and Silicon Graphics.
- * 
- * THE SOFTWARE IS PROVIDED "AS-IS" AND WITHOUT WARRANTY OF ANY KIND, 
- * EXPRESS, IMPLIED OR OTHERWISE, INCLUDING WITHOUT LIMITATION, ANY 
- * WARRANTY OF MERCHANTABILITY OR FITNESS FOR A PARTICULAR PURPOSE.  
- *
- * IN NO EVENT SHALL STANFORD OR SILICON GRAPHICS BE LIABLE FOR
- * ANY SPECIAL, INCIDENTAL, INDIRECT OR CONSEQUENTIAL DAMAGES OF ANY KIND,
- * OR ANY DAMAGES WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS,
- * WHETHER OR NOT ADVISED OF THE POSSIBILITY OF DAMAGE, AND ON ANY THEORY OF 
- * LIABILITY, ARISING OUT OF OR IN CONNECTION WITH THE USE OR PERFORMANCE 
- * OF THIS SOFTWARE.
- */
-
-/*
- * Observable - object to observe
- */
-
-#include <InterViews/observe.h>
-#include "utils/enumerate.h"
-
-Observable::~Observable() {
-	// in case a disconnect removes items from the observers
-    for (long long i = static_cast<long long>(observers_.size()) - 1; i >= 0; --i) {
-	    observers_[i]->disconnect(this);
-        if (i > observers_.size()) {
-            i = observers_.size();
-        }
-	}
-}
-
-void Observable::attach(Observer* o) {
-    observers_.push_back(o);
-}
-
-void Observable::detach(Observer* o) {
-    erase_first(observers_, o);
-}
-
-void Observable::notify() {
-	for (auto& obs: observers_) {
-	    obs->update(this);
-    }
-}
